Fix the user id
Earlier it used to open the id of last user from the list. Even on clicking the first user it used to open last user's id.
Array is used to store the users and position variable is used to point to the exact user from the array.
This implementation can be improved.